Fix grayscale when single `image_ch` selected
Some that have an explicit `num_channels` attr. When it's set with
`image_ch`, grayscale breaks image. When a single `image_ch` is set
output image will have RGB channel with having the same value and
force `num_channels=3`. I know it's waste of memeory and computation,
but it's the best way currently.